+++ test/suites/MouseApiTest.php
@@ -1,0 +1,63 @@
+<?php
+/**
+ * @license see LICENSE
+ */
+
+namespace HeadlessChromium\Test;
+
+use HeadlessChromium\Browser;
+use HeadlessChromium\BrowserFactory;
+
+/**
+ * @covers \HeadlessChromium\Browser
+ * @covers \HeadlessChromium\Page
+ */
+class MouseApiTest extends BaseTestCase
+{
+
+    /**
+     * @var Browser\ProcessAwareBrowser
+     */
+    public static $browser;
+
+    public static function setUpBeforeClass()
+    {
+        parent::setUpBeforeClass();
+        $factory = new BrowserFactory();
+        self::$browser = $factory->createBrowser();
+    }
+
+    public static function tearDownAfterClass()
+    {
+        parent::tearDownAfterClass();
+        self::$browser->close();
+    }
+
+    private function openSitePage($file)
+    {
+        $page = self::$browser->createPage();
+        $page->navigate($this->sitePath($file))->waitForNavigation();
+
+        return $page;
+    }
+
+    /**
+     * @throws \HeadlessChromium\Exception\CommunicationException
+     * @throws \HeadlessChromium\Exception\NoResponseAvailable
+     */
+    public function testClickLink()
+    {
+        // initial navigation
+        $page = $this->openSitePage('b.html');
+        $rect = $page
+            ->evaluate('JSON.parse(JSON.stringify(document.querySelector("#a").getBoundingClientRect()));')
+            ->getReturnValue();
+
+        $page->mouse()->move($rect['x'], $rect['y'])->click();
+        $page->waitForReload();
+
+        $title = $page->evaluate('document.title')->getReturnValue();
+
+        $this->assertEquals('a - test', $title);
+    }
+}